source/lambda/online/requirements.txt
langchain-aws==0.2.17
langgraph==0.2.43
langchain-community==0.3.5
opensearch-py==2.2.0
requests_aws4auth==1.2.2
pydantic==2.11.1
datamodel-code-generator==0.26.2
langchain-openai==0.3.7
cachetools==5.5.2
langchain-mcp-adapters==0.0.9
mcp==1.6.0
fastapi==0.115.12
uvicorn==0.34.2
sse-starlette==2.2.1
nest-asyncio==1.6.0
click==8.1.8
fastapi-mcp==0.3.3
PyJWT==2.8.0
